Fixed the bug

+++ app/src/main/java/hottopic/mit/co/nz/cleaningservice/view/user/LoginActivity.java
@@ -65,7 +65,7 @@ public class LoginActivity extends BaseActivity implements IUserView {
 
             int lastLoginTimestamp = GeneralUtil.getIntFromSP(this, Constants.SP_KEY_LAST_LOGIN_TIMESTAMP);
             // Check last login timestamp, if less than 1 day, do not ask for login
-            if ((int) System.currentTimeMillis() / 1000 - lastLoginTimestamp < MAX_SIGNIN_DURATION) {
+            if (System.currentTimeMillis() / 1000 - lastLoginTimestamp < MAX_SIGNIN_DURATION) {
                 Constants.userInfo = userInfo;
                 ServiceTypes serviceTypes = GeneralUtil.fromJson(GeneralUtil.getDataFromSP(this, Constants.SP_KEY_SERVICE_TYPE), ServiceTypes.class);
                 Intent intent;
